Add version check for SignalBus in NuGet workflows
Introduce a mechanism to check for a specific version of the
`SignalBus` library in the `publish-fluentsignals-signalbus.yml`,
`publish-libs-nuget.yml`, and `publish-nuget.yml` files. If a
version is found in the `FluentSignals.SignalBus.csproj`, it is
used for packaging and pushing to NuGet. Otherwise, the version
from `Directory.Build.props` is used as a fallback, ensuring
the correct version is utilized during the process.